+++ selfdrive/ui/qt/device/device.cc
@@ -1,0 +1,82 @@
+#include "selfdrive/ui/qt/device/device.h"
+
+#include <cmath>
+
+#include <QtConcurrent>
+
+#include "common/swaglog.h"
+
+Device::Device(QObject *parent) : brightness_filter(BACKLIGHT_OFFROAD, BACKLIGHT_TS, BACKLIGHT_DT), QObject(parent) {
+  setAwake(true);
+  resetInteractiveTimeout();
+
+  QObject::connect(uiState(), &UIState::uiUpdate, this, &Device::update);
+}
+
+void Device::update(const UIState &s) {
+  updateBrightness(s);
+  updateWakefulness(s);
+}
+
+void Device::setAwake(bool on) {
+  if (on != awake) {
+    awake = on;
+    Hardware::set_display_power(awake);
+    LOGD("setting display power %d", awake);
+    emit displayPowerChanged(awake);
+  }
+}
+
+void Device::resetInteractiveTimeout(int timeout) {
+  if (timeout == -1) {
+    timeout = (ignition_on ? 10 : 30);
+  }
+  interactive_timeout = timeout * UI_FREQ;
+}
+
+void Device::updateBrightness(const UIState &s) {
+  float clipped_brightness = offroad_brightness;
+  if (s.scene.started && s.scene.light_sensor > 0) {
+    clipped_brightness = s.scene.light_sensor;
+
+    // CIE 1931 - https://www.photonstophotos.net/GeneralTopics/Exposure/Psychometric_Lightness_and_Gamma.htm
+    if (clipped_brightness <= 8) {
+      clipped_brightness = (clipped_brightness / 903.3);
+    } else {
+      clipped_brightness = std::pow((clipped_brightness + 16.0) / 116.0, 3.0);
+    }
+
+    // Scale back to 10% to 100%
+    clipped_brightness = std::clamp(100.0f * clipped_brightness, 10.0f, 100.0f);
+  }
+
+  int brightness = brightness_filter.update(clipped_brightness);
+  if (!awake) {
+    brightness = 0;
+  }
+
+  if (brightness != last_brightness) {
+    if (!brightness_future.isRunning()) {
+      brightness_future = QtConcurrent::run(Hardware::set_brightness, brightness);
+      last_brightness = brightness;
+    }
+  }
+}
+
+void Device::updateWakefulness(const UIState &s) {
+  bool ignition_just_turned_off = !s.scene.ignition && ignition_on;
+  ignition_on = s.scene.ignition;
+
+  if (ignition_just_turned_off) {
+    resetInteractiveTimeout();
+  } else if (interactive_timeout > 0 && --interactive_timeout == 0) {
+    emit interactiveTimeout();
+  }
+
+  setAwake(s.scene.ignition || interactive_timeout > 0);
+}
+
+Device *device() {
+  static Device _device;
+  return &_device;
+}
